sw_engine: --compiler warnings on MSYS2
comparison of integer expressions of different signedness: 'uint32_t' {aka 'unsigned int'} and 'long int' [-Wsign-compare]

+++ src/renderer/sw_engine/tvgSwPostEffect.cpp
@@ -417,8 +417,8 @@ bool effectFill(SwCompositor* cmp, const RenderEffectFill* params, bool direct)
     auto opacity = direct ? MULTIPLY(params->color[3], cmp->opacity) : params->color[3];
 
     auto& bbox = cmp->bbox;
-    auto w = (bbox.max.x - bbox.min.x);
-    auto h = (bbox.max.y - bbox.min.y);
+    auto w = size_t(bbox.max.x - bbox.min.x);
+    auto h = size_t(bbox.max.y - bbox.min.y);
     auto color = cmp->recoverSfc->join(params->color[0], params->color[1], params->color[2], 255);
 
     TVGLOG("SW_ENGINE", "Fill region(%ld, %ld, %ld, %ld), param(%d %d %d %d)", bbox.min.x, bbox.min.y, bbox.max.x, bbox.max.y, params->color[0], params->color[1], params->color[2], params->color[3]);
@@ -426,10 +426,10 @@ bool effectFill(SwCompositor* cmp, const RenderEffectFill* params, bool direct)
     if (direct) {
         auto dbuffer = cmp->recoverSfc->buf32 + (bbox.min.y * cmp->recoverSfc->stride + bbox.min.x);
         auto sbuffer = cmp->image.buf32 + (bbox.min.y * cmp->image.stride + bbox.min.x);
-        for (uint32_t y = 0; y < h; ++y) {
+        for (size_t y = 0; y < h; ++y) {
             auto dst = dbuffer;
             auto src = sbuffer;
-            for (uint32_t x = 0; x < w; ++x, ++dst, ++src) {
+            for (size_t x = 0; x < w; ++x, ++dst, ++src) {
                 auto a = MULTIPLY(opacity, A(*src));
                 auto tmp = ALPHA_BLEND(color, a);
                 *dst = tmp + ALPHA_BLEND(*dst, 255 - a);
@@ -440,9 +440,9 @@ bool effectFill(SwCompositor* cmp, const RenderEffectFill* params, bool direct)
         cmp->valid = true;  //no need the subsequent composition
     } else {
         auto dbuffer = cmp->image.buf32 + (bbox.min.y * cmp->image.stride + bbox.min.x);
-        for (uint32_t y = 0; y < h; ++y) {
+        for (size_t y = 0; y < h; ++y) {
             auto dst = dbuffer;
-            for (uint32_t x = 0; x < w; ++x, ++dst) {
+            for (size_t x = 0; x < w; ++x, ++dst) {
                 *dst = ALPHA_BLEND(color, MULTIPLY(opacity, A(*dst)));
             }
             dbuffer += cmp->image.stride;
